# apps/server/src/static/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"DefaultSelection_Prisma._36_userPayload_": {
				"properties": {
					"salt": {
						"type": "string"
					},
					"avatar": {
						"type": "string"
					},
					"nickname": {
						"type": "string"
					},
					"des": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"role": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"salt",
					"avatar",
					"nickname",
					"des",
					"username",
					"role",
					"password",
					"email",
					"id"
				],
				"type": "object"
			},
			"User": {
				"$ref": "#/components/schemas/DefaultSelection_Prisma._36_userPayload_",
				"description": "Model user"
			},
...
			"JsonResultType__token-string--me-User__": {
				"properties": {
					"code": {
						"type": "number",
						"format": "double"
					},
					"msg": {
						"type": "string"
					},
					"data": {
						"properties": {
							"me": {
								"$ref": "#/components/schemas/User"
							},
							"token": {
								"type": "string"
							}
						},
						"required": [
							"me",
							"token"
						],
						"type": "object"
					}
				},
				"required": [
					"code",
					"msg",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse__token-string--me-User__": {
				"$ref": "#/components/schemas/JsonResultType__token-string--me-User__",
				"description": "控制器方法返回类型的通用别名\r\n用于简化 Promise<JsonResultType<T>> 的写法"
			},
			"LoginParams": {
				"properties": {
					"aseKeyEncrypt": {
						"type": "string"
					},
					"code": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"account": {
						"type": "string"
					}
				},
				"type": "object"
			},
			"JsonResultType_any_": {
				"properties": {
					"code": {
						"type": "number",
						"format": "double"
					},
					"msg": {
						"type": "string"
					},
					"data": {}
				},
				"required": [
					"code",
					"msg",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse": {
				"$ref": "#/components/schemas/JsonResultType_any_",
				"description": "控制器方法返回类型的通用别名\r\n用于简化 Promise<JsonResultType<T>> 的写法"
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "Server API",
		"version": "1.0.0",
		"description": "## Git Drive 接口文档\n\n  这是一个基于 Koa + TSOA 的后端 API 服务",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"paths": {
...
		"/api/login": {
			"post": {
				"operationId": "RootController-Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse__token-string--me-User__"
								}
							}
						}
					}
				},
				"description": "支持多种登录方式。\r\n首次登录请用邮箱验证码登录方式，进去之后再设置密码，以后便可以用账密方式登录了！\r\n\r\n| 登录方式 | 字段 | 说明 |\r\n|----------|----------|------|\r\n| 账号密码 | username, password | 账密登录 |\r\n| 邮箱验证码 | email, code | 邮箱 + 验证码登录 |",
				"summary": "登录",
				"tags": [
					"root"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginParams"
							}
						}
					}
				}
			}
		},
		"/api/logout": {
			"post": {
				"operationId": "RootController-Logout",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "退出登录",
				"summary": "登出",
				"tags": [
					"root"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/send-code": {
			"post": {
				"operationId": "RootController-SendCode",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "发送证码 接口",
				"summary": "发送证码",
				"tags": [
					"root"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"type": {
										"type": "string"
									},
									"phone": {
										"type": "string"
									},
									"email": {
										"type": "string"
									}
								},
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/send-reset-pwd-link": {
			"post": {
				"operationId": "RootController-SendResetPwdLink",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "发送重置密码链接 接口",
				"summary": "发送重置密码链接",
				"tags": [
					"root"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"email": {
										"type": "string"
									}
								},
								"required": [
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/reset-pwd": {
			"post": {
				"operationId": "RootController-ResetPwd",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"summary": "重置密码",
				"tags": [
					"root"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"code": {
										"type": "string"
									},
									"email": {
										"type": "string"
									},
									"aseKeyEncrypt": {
										"type": "string"
									},
									"password": {
										"type": "string"
									}
								},
								"required": [
									"code",
									"email",
									"aseKeyEncrypt",
									"password"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	],
...
}